move TemplatingData to separate pkg
