+++ scripts/image-builder/configure.sh
@@ -29,6 +29,8 @@ if grep -qE "Red Hat Enterprise Linux.*Beta" /etc/redhat-release; then
 fi
 
 sudo systemctl enable osbuild-composer.socket --now
+sudo usermod -a -G weldr "$(whoami)"
+
 sudo systemctl enable cockpit.socket --now
 sudo firewall-cmd --add-service=cockpit --permanent
