+++ test/pkg/compose/templatingdata/assembler.go
@@ -1,4 +1,4 @@
-package compose
+package templatingdata
 
 import (
 	"encoding/json"
@@ -24,67 +24,12 @@ var (
 	errNoRemoteRelease = fmt.Errorf("release from remote repo not found")
 )
 
-// Release represents metadata of particular set of RPMs.
-type Release struct {
-	// Repository is where RPM resides. It can be local (on the disk), http (like OpenShift's mirror), or RHOCP.
-	Repository string
-
-	// Version is full version string of a RPM, e.g. 4.14.16-202403071942.p0.g4cef5f2.assembly.4.14.16.el9.
-	Version string
-
-	// Minor is minor part of the version, e.g. 14.
-	Minor int
-
-	// Images is a list of images stored in release-info RPM.
-	// Currently only for local repositories.
-	// TODO: Extend to remote for local image mirroring.
-	Images []string
+type TemplatingDataOpts struct {
+	ArtifactsMainDir               string
+	TemplatingDataFragmentFilepath string
 }
 
-// TemplatingData contains all values needed for templating Composer's Sources & Blueprints,
-// and other templated artifacts within a MicroShift's test harness.
-type TemplatingData struct {
-	Arch string
-
-	// Minor version of current release's RHOCP. If RHOCP is not available yet, it defaults to 0.
-	RHOCPMinorY int
-
-	// Minor version of previous release's RHOCP. If RHOCP is not available yet, it defaults to 0.
-	RHOCPMinorY1 int
-
-	// Minor version of previous previous release's RHOCP.
-	RHOCPMinorY2 int
-
-	// Current stores metadata of current release, i.e. matching currently checked out git branch.
-	// If the RHOCP is not available yet, it can point to Release or Engineering candidates present on the OpenShift mirror.
-	// If those are also not available, it will be empty and related composer Sources and Blueprints will not be build.
-	Current Release
-
-	// Current stores metadata of current release, i.e. matching currently checked out git branch.
-	// If the RHOCP is not available yet, it can point to Release or Engineering candidates present on the OpenShift mirror.
-	Previous Release
-
-	// Current stores metadata of current release, i.e. matching currently checked out git branch.
-	// Usually this should always point to RHOCP because it's two minor versions older than what we're working on currently.
-	YMinus2 Release
-
-	// Source stores metadata of RPMs built from currently checked out source code.
-	Source Release
-
-	// Source stores metadata of RPMs built from base branch of currently checked out branch.
-	// Usually it can be `main` or `release-4.Y`.
-	Base Release
-
-	// Source stores metadata of RPMs built from currently checked out source code with minor version overridden
-	// to be newer than what we're currently working on.
-	// These are needed for various ostree upgrade tests.
-	FakeNext Release
-
-	// External stores metadata of RPMs supplied from external source, like private builds.
-	External Release
-}
-
-func NewTemplatingData(opts *ComposeOpts) (*TemplatingData, error) {
+func New(opts *TemplatingDataOpts) (*TemplatingData, error) {
 	rpmRepos := path.Join(opts.ArtifactsMainDir, "rpm-repos")
 	localRepo := path.Join(rpmRepos, "microshift-local")
 	fakeNextRepo := path.Join(rpmRepos, "microshift-fake-next-minor")
@@ -296,7 +241,7 @@ func getReleaseFromTheMirror(minor int, devPreview bool) (Release, error) {
 		dp = "-dev-preview"
 	}
 
-	repo := fmt.Sprintf("https://mirror.openshift.com/pub/openshift-v4/%s/microshift/ocp%s/latest-4.%d/el9/os", getArch(), dp, minor)
+	repo := fmt.Sprintf("https://mirror.openshift.com/pub/openshift-v4/%s/microshift/ocp%s/latest-4.%d/el9/os/", getArch(), dp, minor)
 
 	resp, err := http.Get(repo)
 	if err != nil {
